# ref(profiling): text renderer perf