--- a/static/app/utils/profiling/renderers/textRenderer.tsx
+++ b/static/app/utils/profiling/renderers/textRenderer.tsx
@@ -12,15 +12,6 @@ import {
   trimTextCenter,
 } from '../gl/utils';
 
-export function isOutsideView(frame: Rect, view: Rect): boolean {
-  // Frame is outside of the view on the left
-  if (frame.overlaps(view)) {
-    return false;
-  }
-
-  return true;
-}
-
 class TextRenderer {
   textCache: Record<string, number> = {};
 
@@ -72,27 +63,16 @@ class TextRenderer {
 
     const minWidth = this.measureAndCacheText(ELLIPSIS);
 
-    let frame: FlamegraphFrame;
-    let i: number;
-    const length: number = this.flamegraph.frames.length;
+    const SIDE_PADDING = 2 * this.theme.SIZES.BAR_PADDING * window.devicePixelRatio;
+    const HALF_SIDE_PADDING = SIDE_PADDING / 2;
+    const BASELINE_OFFSET =
+      (this.theme.SIZES.BAR_HEIGHT - this.theme.SIZES.BAR_FONT_SIZE / 2) *
+      window.devicePixelRatio;
 
-    // We currently iterate over all frames, but we could optimize this to only iterate over visible frames.
-    // This could be achieved by querying the flamegraph tree (as an interval tree). This would still run in O(n), but
-    // would improve our best case performance (e.g. when users) zoom into the flamegraph
-    for (i = 0; i < length; i++) {
-      frame = this.flamegraph.frames[i];
-
-      // This rect gets discarded after each render which is wasteful
-      const frameInConfigSpace = new Rect(
-        frame.start,
-        this.flamegraph.inverted ? configSpace.height - frame.depth - 1 : frame.depth,
-        frame.end - frame.start,
-        1
-      );
-
+    const drawFrame = (frame: FlamegraphFrame): void => {
       // Check if our rect overlaps with the current viewport and skip it
-      if (isOutsideView(frameInConfigSpace, configViewSpace)) {
-        continue;
+      if (frame.end <= configViewSpace.left || frame.start >= configViewSpace.right) {
+        return;
       }
 
       // We pin the start and end of the frame, so scrolling around keeps text pinned to the left or right side of the viewport
@@ -102,7 +82,7 @@ class TextRenderer {
       // This rect gets discarded after each render which is wasteful
       const offsetFrame = new Rect(
         pinnedStart,
-        frameInConfigSpace.y,
+        this.flamegraph.inverted ? configSpace.height - frame.depth - 1 : frame.depth,
         pinnedEnd - pinnedStart,
         1
       );
@@ -112,38 +92,44 @@ class TextRenderer {
 
       // Since the text is not exactly aligned to the left/right bounds of the frame, we need to subtract the padding
       // from the total width, so that we can truncate the center of the text accurately.
-      const paddedRectangleWidth =
-        frameInPhysicalSpace.width -
-        2 * this.theme.SIZES.BAR_PADDING * window.devicePixelRatio;
+      const paddedRectangleWidth = frameInPhysicalSpace.width - SIDE_PADDING;
 
       // We want to draw the text in the vertical center of the frame, so we substract half the height of the text
-      const y =
-        frameInPhysicalSpace.y +
-        (this.theme.SIZES.BAR_HEIGHT - this.theme.SIZES.BAR_FONT_SIZE / 2) *
-          window.devicePixelRatio;
+      const y = frameInPhysicalSpace.y + BASELINE_OFFSET;
 
       // Offset x by 1x the padding
-      const x =
-        frameInPhysicalSpace.x + this.theme.SIZES.BAR_PADDING * window.devicePixelRatio;
+      const x = frameInPhysicalSpace.x + HALF_SIDE_PADDING;
 
       // If the width of the text is greater than the minimum width to render, we should render it
-      if (paddedRectangleWidth >= minWidth) {
-        let text = frame.frame.name;
-
-        // If text width is smaller than rectangle, just draw the text
-        if (this.measureAndCacheText(text) > paddedRectangleWidth) {
-          text = trimTextCenter(
-            text,
-            findRangeBinarySearch(
-              {low: 0, high: paddedRectangleWidth},
-              n => this.measureAndCacheText(text.substring(0, n)),
-              paddedRectangleWidth
-            )[0]
-          );
-        }
-
-        this.context.fillText(text, x, y);
+      // Since children of a frame cannot be wider than the frame itself, we can exit early and discard the entire subtree
+      if (paddedRectangleWidth <= minWidth) {
+        return;
       }
+
+      this.context.fillText(
+        trimTextCenter(
+          frame.frame.name,
+          findRangeBinarySearch(
+            {low: 0, high: paddedRectangleWidth},
+            n => this.measureAndCacheText(frame.frame.name.substring(0, n)),
+            paddedRectangleWidth
+          )[0]
+        ),
+        x,
+        y
+      );
+
+      for (let i = 0; i < frame.children.length; i++) {
+        drawFrame(frame.children[i]);
+      }
+    };
+
+    // We start by iterating over root frames, so we draw the call stacks top-down.
+    // This allows us to do a couple optimizations that improve our best case performance.
+    // 1. We can skip drawing the entire tree if the root frame is not visible
+    // 2. We can skip drawing and
+    for (let i = 0; i < this.flamegraph.roots.length; i++) {
+      drawFrame(this.flamegraph.roots[i]);
     }
   }
 }