Improve TextRenderer performance with a few smarter heuristics and remove a few branches. Since we now draw the tree top down, we can eliminate entire parts of the subtree faster. It turns out that allocating a new Rect per frame and per draw call that gets immediately gc'd is also costly. We can eliminate the first rect by just computing the position on our own, which boils down the whole "in view" check to a simple math operation.

**Before:**
typescript (right quarter) x 845 ops/sec ±1.23% (92 runs sampled)
typescript (center half) x 717 ops/sec ±2.17% (87 runs sampled)
typescript (full profile) x 375 ops/sec ±1.14% (92 runs sampled)

**After:**
typescript (right quarter) x 7,361 ops/sec ±3.14% (92 runs sampled)
typescript (center half) x 5,174 ops/sec ±2.64% (92 runs sampled)
typescript (full profile) x 1,143 ops/sec ±2.29% (96 runs sampled)